fix(ruby-core-lib): fix returning nil instead of empty request params (#17)
The faraday-client auto-sets the content-type for the empty body and it does not set the content-type for nil body. From now onwards, the default body will be nil if the body is not associated with the outgoing request. This commit removes the unless block and takes out the block's body.